+++ CHANGES.rst
@@ -7,7 +7,10 @@ Changelog
 
 - Dont double decode XML HTML pages.
   [rofersou]
+
 - Pass headers forward from original request.
+  [rodfersou]
+
 - Make URL field not required.
   [rodfersou]