+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.2.2 (unreleased)
+1.2.2 (2019-05-28)
 ------------------
 
 - Dont double decode XML HTML pages.